verifica se o id existe no bd

+++ forum/src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -2,6 +2,7 @@ package br.com.alura.forum.controller;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
@@ -40,11 +41,11 @@ public class TopicosController {
 //	@ResponseBody //navega pelo browser sem precisar criar uma página.html
 	public List<TopicoDto> lista(String nomeCurso) {
 //		System.out.println(nomeCurso);
-		if(nomeCurso == null) {
+		if(nomeCurso == null) { //caso o nome for null irá trazer todos
 			List<Topico> topicos = topicoRepository.findAll();
 			return TopicoDto.converter(topicos);
 			
-		} else {
+		} else { //caso especificar um nome, irá trazer a lista que tem aquele nome
 			List<Topico> topicos = topicoRepository.findByCursoNome(nomeCurso);
 			return TopicoDto.converter(topicos);
 		}
@@ -67,25 +68,33 @@ public class TopicosController {
 	 * detalhar um tópico especifico
 	 */
 	@GetMapping("/{id}")
-	public DetalhesDoTopicoDto detalhar(@PathVariable Long id) {
-		Topico topico = topicoRepository.getOne(id);
-		return new DetalhesDoTopicoDto(topico);
-			
+	public ResponseEntity<DetalhesDoTopicoDto> detalhar(@PathVariable Long id) {
+		//getOne sempre irá considerar que o id existe
+		Optional<Topico> topico = topicoRepository.findById(id);
+		if(topico.isPresent()) { //verifica se o id está presente no banco de dados
+			return ResponseEntity.ok(new DetalhesDoTopicoDto(topico.get()));
+		}	
+		return ResponseEntity.notFound().build();
 	}
 	
 	@PutMapping("/{id}") //sobreescrever o recurso (algo que já está escrito)
 	@Transactional //informar que é para commitar a nova alteração no banco.
 	public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form){
-		Topico topico = form.atualizar(id, topicoRepository);
-		
-		return ResponseEntity.ok(new TopicoDto(topico));
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if(optional.isPresent()) { //verifica se o id está presente no banco de dados
+			Topico topico = form.atualizar(id, topicoRepository);
+			return ResponseEntity.ok(new TopicoDto(topico));
+		}	
+		return ResponseEntity.notFound().build(); //irá retornar erro 404 caso não encontre nenhum id
 	}
 	
 	@DeleteMapping("{id}") //deletar um topico especifico
 	public ResponseEntity<?> remover(@PathVariable Long id){ //<?> generico sem especificar um
-		topicoRepository.deleteById(id);
-		
-		return ResponseEntity.ok().build(); //apenas irá retornar o retorno ok após o topico ser deletado
-		
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if(optional.isPresent()) { 
+			topicoRepository.deleteById(id);
+			return ResponseEntity.ok().build(); //apenas irá retornar o retorno ok após o topico ser deletado
+		}
+		return ResponseEntity.notFound().build();
 	}
 }
